Feature: Add global filtering capabilities (#735)
* Make filter components dumb and export them

* Add tabCookies prop to landing page section

* Break landing page component and add filters

* Add overflow to pages

* Adjust height of landing page section container

* Break landing page into small component and add filter to dashboard

* Add key into filtering hook

* Implement query in sitemap report and pass as props

* Remove top border for cookie section

* Memo the cookies and calculate persistence key earlier

* Pass sidebar key rather than calculating again

* Make category filter static and update tests

* Add exemption reason filter and pass data to report from ext

* Fix lint

* Add filter and update use mergeDeep for storage

* Add filtered data and filters to report object for a site

* Add code to download filtered report from sitemap html

* Update tests

* Update height of panel for toastMessage

* Fix lint action

* Fix lint action

* Calculate cookies with issues inside landing page and remove props

* Pass libraryMatches as prop

* Make category filters static

* Load dummy data and hot reload dashboard in dev mode

* Read dummy data

* Update states and side effects

* Update dev command

* Pass query to component and use in filters

* Create seperate filter hook, create cookies tab component for pre-processing query

* Move filtering hook to landing page

* Add publicpath

* Allow passing a frame of choice for navigation

* Update tests

* Fix report callbacks

* move global filter hook to design system, change key and use hook in ext

* Add tab selection to exemption reason

* use correct imports

* Silent React import warning in library detection

* Remove filters from extension download

* Add correct imports

* Pass unfiltered data to dashboard report downloaded from ext

* don't pass pageurl to site cookies

* Update flags

* Remove comments and move dependecies

---------

+++ packages/cli-dashboard/src/app.tsx
@@ -16,7 +16,7 @@
 /**
  * External dependencies
  */
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import type {
   CompleteJson,
   CookieFrameStorageType,
@@ -53,13 +53,6 @@ const App = () => {
     [key: string]: LibraryData;
   } | null>(null);
 
-  const type = useMemo(() => {
-    // @ts-ignore
-    return globalThis?.PSAT_DATA?.type === 'sitemap'
-      ? DisplayType.SITEMAP
-      : DisplayType.SITE;
-  }, []);
-
   useEffect(() => {
     const bodyTag = document.querySelector('body');
 
@@ -70,7 +63,29 @@ const App = () => {
     bodyTag.style.fontSize = '75%';
   }, []);
 
+  const [type, setType] = useState<DisplayType>(DisplayType.SITE);
+
   useEffect(() => {
+    (async () => {
+      if (process.env.NODE_ENV === 'development') {
+        const module = await import('./dummyData/PSAT_DATA.js');
+        // @ts-ignore
+        globalThis.PSAT_DATA = module.default;
+      }
+
+      setType(
+        globalThis?.PSAT_DATA?.type === 'sitemap'
+          ? DisplayType.SITEMAP
+          : DisplayType.SITE
+      );
+    })();
+  }, []);
+
+  useEffect(() => {
+    if (!globalThis?.PSAT_DATA) {
+      return;
+    }
+
     sessionStorage.clear();
     //@ts-ignore
     const messages = globalThis?.PSAT_DATA?.translations;
@@ -92,7 +107,7 @@ const App = () => {
       _libraryMatches = extractedData.consolidatedLibraryMatches;
       setLandingPageCookies(extractedData.landingPageCookies);
     } else {
-      _cookies = extractCookies(data[0].cookieData, data[0].pageUrl, true);
+      _cookies = extractCookies(data[0].cookieData, '', true);
       _technologies = data[0].technologyData;
       _libraryMatches = { [data[0].pageUrl]: data[0].libraryMatches };
     }
